requirements.txt
Django==3.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
django-filter==21.1
djangorestframework==3.12.4
